requirements.txt
fastapi
uvicorn
psycopg2
python-dotenv
sqlalchemy
alembic
schemas
pydantic
PyJWT
pydantic[email]
passlib
passlib[bcrypt]
fastapi_mail
jinja2